Create ArtistClass.dart and ArtworkClass.dart
