+++ jacoby_arts/lib/auxiliary/CartClasses.dart
@@ -2,11 +2,13 @@
 class ArtInfo {
   String artworkName;
   String artistName;
-  double price;
+  int price;
+  String imageUrl;
 
-  ArtInfo(String artworkName, String artistName, double price) {
+  ArtInfo(String artworkName, String artistName, int price, String imageUrl) {
     this.artworkName = artworkName;
     this.artistName = artistName;
     this.price = price;
+    this.imageUrl = imageUrl;
   }
 }
