+++ jacoby_arts/lib/widgets/ArtworkListPageBody.dart
@@ -1,64 +1,65 @@
 import 'package:flutter/material.dart';
 import 'package:jacoby_arts/pages/ArtworkDetailsPage.dart';
-import 'package:jacoby_arts/widgets/ArtworkListPageBody.dart';
+import 'package:jacoby_arts/auxiliary/ArtistClass.dart';
+import 'package:jacoby_arts/auxiliary/ArtworkClass.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:jacoby_arts/Auxiliary/uiComponents.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'dart:async';
 
 
-class Artist{
-  final String artist_name;
-  final DocumentReference reference;
+// class Artist{
+//   final String artist_name;
+//   final DocumentReference reference;
 
-  Artist.fromMap(Map<String,dynamic> map, {this.reference})
-  :assert(map['name'] != null),
-  artist_name = map['name'];
+//   Artist.fromMap(Map<String,dynamic> map, {this.reference})
+//   :assert(map['name'] != null),
+//   artist_name = map['name'];
 
-  Artist.fromSnapshot(DocumentSnapshot snapshot)
-  :this.fromMap(snapshot.data, reference: snapshot.reference);
+//   Artist.fromSnapshot(DocumentSnapshot snapshot)
+//   :this.fromMap(snapshot.data, reference: snapshot.reference);
 
-  @override
-  String toString() => "Artist<$artist_name>";
-}
+//   @override
+//   String toString() => "Artist<$artist_name>";
+// }
 
-class Artwork{
-  final String artist_id;
-  //final String artwork_id;
-  final String cart_id;
-  final String description;
-  final String exhibit_name;
-  final String image_url;
-  final String market_status;
-  final int price;
-  final String title;
-  final DocumentReference reference;
+// class Artwork{
+//   final String artist_id;
+//   //final String artwork_id;
+//   final String cart_id;
+//   final String description;
+//   final String exhibit_name;
+//   final String image_url;
+//   final String market_status;
+//   final int price;
+//   final String title;
+//   final DocumentReference reference;
 
-  Artwork.fromMap(Map<String, dynamic> map, {this.reference})
-    :assert(map['artist_id'] != null),
-   // assert(map['artwork_id'] != null),
-    assert(map['cart_id'] != null),
-    assert(map['description'] != null),
-    assert(map['exhibit_name'] != null),
-    assert(map['image_URL'] != null),
-    assert(map['market_status'] != null),
-    assert(map['price'] != null),
-    assert(map['title'] != null),
-    artist_id = map['artist_id'],
-    //artwork_id = map['artwork_id'],
-    cart_id = map['cart_id'],
-    description = map['description'],
-    exhibit_name = map['exhibit_name'],
-    image_url = map['image_URL'],
-    market_status = map['market_status'],
-    price = map['price'],
-    title = map['title'];
-  Artwork.fromSnapshot(DocumentSnapshot snapshot)
-    :this.fromMap(snapshot.data, reference: snapshot.reference);
-  @override
-  String toString() => "ExhibitName<$artist_id:$cart_id:$description:$exhibit_name:$image_url:$market_status:$price:$title>";
-  //$artwork_id:
-}
+//   Artwork.fromMap(Map<String, dynamic> map, {this.reference})
+//     :assert(map['artist_id'] != null),
+//    // assert(map['artwork_id'] != null),
+//     assert(map['cart_id'] != null),
+//     assert(map['description'] != null),
+//     assert(map['exhibit_name'] != null),
+//     assert(map['image_URL'] != null),
+//     assert(map['market_status'] != null),
+//     assert(map['price'] != null),
+//     assert(map['title'] != null),
+//     artist_id = map['artist_id'],
+//     //artwork_id = map['artwork_id'],
+//     cart_id = map['cart_id'],
+//     description = map['description'],
+//     exhibit_name = map['exhibit_name'],
+//     image_url = map['image_URL'],
+//     market_status = map['market_status'],
+//     price = map['price'],
+//     title = map['title'];
+//   Artwork.fromSnapshot(DocumentSnapshot snapshot)
+//     :this.fromMap(snapshot.data, reference: snapshot.reference);
+//   @override
+//   String toString() => "ExhibitName<$artist_id:$cart_id:$description:$exhibit_name:$image_url:$market_status:$price:$title>";
+//   //$artwork_id:
+// }
 
 class ArtworkListPageBody extends StatelessWidget {
   ArtworkListPageBody();
@@ -120,47 +121,48 @@ _artist = Artist.fromSnapshot(data);
 InkWell makeCard(BuildContext context, DocumentSnapshot data){
   final _artwork =  Artwork.fromSnapshot(data);
   final _artist =_artwork.artist_id;
-  
-  
-  return new InkWell(
-      // make it clickable
-      onTap: () {
-        // create a event page if its clicked
-        Navigator.push(context,
-        MaterialPageRoute(
-            builder: (__) => new ArtworkDetailsPage(artData: _artwork)
+  var one = 1;
+  if (one == 1){
+    return new InkWell(
+        // make it clickable
+        onTap: () {
+          // create a event page if its clicked
+          Navigator.push(context,
+          MaterialPageRoute(
+              builder: (__) => new ArtworkDetailsPage(artData: _artwork)
 
-          )
-          );
-      },
-      child: Card(
-        elevation: 8.0,
-        margin: new EdgeInsets.symmetric(horizontal: horizontalPadding, vertical: cardPadding),
-        child: Container(
-          decoration: BoxDecoration(color: Colors.blueGrey),
-          child: ListTile(
-            contentPadding: EdgeInsets.symmetric(horizontal: 20.0, vertical: 10.0),
-            leading: Container(
-              padding: EdgeInsets.only(right: 12.0),
-              decoration: new BoxDecoration(
-                 // create a inset for the image
-                border: new Border(
-                right: new BorderSide(width: 1.0, color: Colors.white))),
-                child: new Image.network(_artwork.image_url, height: 75, width: 75,), //url here
-            ),
-              title: Text(
-                _artwork.title,
-                style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+            )
+            );
+        },
+        child: Card(
+          elevation: 8.0,
+          margin: new EdgeInsets.symmetric(horizontal: horizontalPadding, vertical: cardPadding),
+          child: Container(
+            decoration: BoxDecoration(color: Colors.blueGrey),
+            child: ListTile(
+              contentPadding: EdgeInsets.symmetric(horizontal: 20.0, vertical: 10.0),
+              leading: Container(
+                padding: EdgeInsets.only(right: 12.0),
+                decoration: new BoxDecoration(
+                  // create a inset for the image
+                  border: new Border(
+                  right: new BorderSide(width: 1.0, color: Colors.white))),
+                  child: new Image.network(_artwork.image_url, height: 75, width: 75,), //url here
               ),
-              subtitle: Row(
-                children: <Widget>[
-                  
-                  Text(_artist, style: TextStyle(color: Colors.white)),
-                  Text("\$" + _artwork.price.toString() + ".00 ", style: TextStyle(color: Colors.white)),
-                ],
-              ),
-              trailing:
-                Icon(Icons.keyboard_arrow_right, color: Colors.white, size: 30.0)),
-        ),
-  ));
+                title: Text(
+                  _artwork.title,
+                  style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+                ),
+                subtitle: Row(
+                  children: <Widget>[
+                    
+                    Text(_artist, style: TextStyle(color: Colors.white)),
+                    Text("\$" + _artwork.price.toString() + ".00 ", style: TextStyle(color: Colors.white)),
+                  ],
+                ),
+                trailing:
+                  Icon(Icons.keyboard_arrow_right, color: Colors.white, size: 30.0)),
+          ),
+    ));
+  }
 }
